implemented separate netApy and borrow costs comparison for refinance
